-아틀라스슬라이서가 타일의 texcoord를 id값으로 저장해주고 확인할수 있도록 해당 타일을 직접 잘라서 png파일로 잘라준다. -tile view, model 작업중이다.

+++ SFML-AtlasSlicer/TargetRect.cpp
@@ -55,10 +55,10 @@ void TargetRect::SliceRect(int x, int y)
 	{
 		for (int i = 0; i < x; i++)
 		{
-			m_SubRect[j][i]->SetFloatRect(sf::FloatRect(m_Size.getPosition().x + i * (m_Size.width /x),
-				m_Size.getPosition().y + j * (m_Size.height / y), 
-				m_Size.width / x, 
-				m_Size.height / y));
+			m_SubRect[j][i]->SetFloatRect(sf::FloatRect(m_Size.getPosition().x + (float)i * (m_Size.width / (float)x),
+				m_Size.getPosition().y + (float)j * (m_Size.height / (float)y),
+				m_Size.width / (float)x,
+				m_Size.height / (float)y));
 		}
 	}
 }
